requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
Django==4.1.2
django-crispy-forms==2.0
django-extensions==3.2.1
djangorestframework==3.14.0
idna==3.4
mysql-connector-python==8.0.32
mysqlclient==2.1.1
protobuf==3.20.3
python-dotenv==1.0.0
pytz==2023.3
razorpay==1.3.0
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.15
django-allauth == 0.54.0
social-auth-app-django == 5.2.0


